HZN-1202: Moved several classes to fix some OSGi loading issues
