+++ core/lib/src/main/java/org/opennms/core/collections/JdbcSet.java
@@ -26,13 +26,15 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.core.utils;
+package org.opennms.core.collections;
 
 import java.util.AbstractSet;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.Set;
+
+import org.opennms.core.utils.IteratorIterator;
 
 
 /**
