+++ opennms-config-jaxb/src/main/java/org/opennms/netmgt/config/poller/Parameter.java
@@ -37,8 +37,6 @@ import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
-import org.opennms.core.xml.JaxbClassObjectAdapter;
-
 
 /**
  * Parameters to be used for polling this service. E.g.: for polling HTTP, the
@@ -67,7 +65,7 @@ public class Parameter implements Serializable {
      * Field m_contents.
      */
     @XmlAnyElement(lax=false)
-    @XmlJavaTypeAdapter(JaxbClassObjectAdapter.class)
+    @XmlJavaTypeAdapter(PollerClassObjectAdapter.class)
     private Object m_contents;
