+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/support/GenericIndexResourceType.java
@@ -41,7 +41,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import org.opennms.core.utils.LazySet;
+import org.opennms.core.collections.LazySet;
 import org.opennms.core.utils.PropertiesUtils;
 import org.opennms.core.utils.PropertiesUtils.SymbolTable;
 import org.opennms.netmgt.collection.api.StorageStrategy;
