+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/support/LazyChildResourceLoader.java
@@ -32,7 +32,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.opennms.core.utils.LazyList;
+import org.opennms.core.collections.LazyList;
 import org.opennms.netmgt.dao.api.ResourceDao;
 import org.opennms.netmgt.model.OnmsResource;
 import org.opennms.netmgt.model.OnmsResourceType;
