+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/support/LazyResourceAttributeLoader.java
@@ -30,7 +30,7 @@ package org.opennms.netmgt.dao.support;
 
 import java.util.Set;
 
-import org.opennms.core.utils.LazySet;
+import org.opennms.core.collections.LazySet;
 import org.opennms.netmgt.dao.api.ResourceStorageDao;
 import org.opennms.netmgt.model.OnmsAttribute;
 import org.opennms.netmgt.model.ResourcePath;
