+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/support/NodeResourceType.java
@@ -34,7 +34,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.opennms.core.utils.LazyList;
+import org.opennms.core.collections.LazyList;
 import org.opennms.netmgt.dao.api.NodeDao;
 import org.opennms.netmgt.dao.api.ResourceDao;
 import org.opennms.netmgt.model.OnmsAttribute;
