+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/support/NodeSnmpResourceType.java
@@ -32,7 +32,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
-import org.opennms.core.utils.LazySet;
+import org.opennms.core.collections.LazySet;
 import org.opennms.netmgt.dao.api.ResourceStorageDao;
 import org.opennms.netmgt.model.OnmsAttribute;
 import org.opennms.netmgt.model.OnmsResource;
