+++ opennms-services/src/main/java/org/opennms/netmgt/passive/PassiveStatusKey.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.config;
+package org.opennms.netmgt.passive;
 
 /**
  * <p>PassiveStatusKey class.</p>
